feat: add searchAgain functionality to QuestionForm
The `QuestionForm` component in the `management_client/src/components/questions/base/QuestionForm.jsx` file has been updated to include the `searchAgain` prop. This prop allows users to perform another search for a question by clicking the "Buscar Otra Pregunta" button. This feature enhances the functionality of the application and improves the user experience.