feat: add delete functionality for questions
This commit adds the `handleDeleteFromList` function to the `formHandlers.js` file in the `delete_questions` directory. This function is responsible for deleting a question from the list of questions. It filters out the question with the specified code and updates the state with the updated list of questions. It also logs a success message and sets the success message state variable.