feat: add SuggestionsList component for displaying suggestions
This commit adds a new component called SuggestionsList, which is responsible for rendering a list of suggestions. It receives an array of suggestions as props and renders them as list items. Each suggestion item is clickable and triggers an onSuggestionClick event when clicked. This component will be used in the `modify_questions` module for displaying suggestions to the user.