feat: add file input support to InputField component
The InputField component has been updated to include support for file inputs. This allows users to upload files when filling out forms. The new fileInput function has been added to handle the rendering of the file input element. Multiple files can be selected if the field allows it. This change improves the flexibility and usability of the InputField component.